+++ lib/karafka/connection/consumer.rb
@@ -3,7 +3,7 @@
 module Karafka
   module Connection
     # Class that consumes messages for which we listen
-    class Consumer
+    module Consumer
       # Consumes a message (does something with it)
       # It will execute a scheduling task from a proper controller based on a message topic
       # @note This should be looped to obtain a constant listening
@@ -11,7 +11,7 @@ module Karafka
       #   for a given consumption will affect other consumed messages
       #   If we would't catch it, it would propagate up until killing the Celluloid actor
       # @param kafka_message [Kafka::FetchedMessage] raw message that was fetched by kafka
-      def consume(kafka_message)
+      def self.consume(kafka_message)
         # We map from incoming topic name, as it might be namespaced, etc.
         # @see topic_mapper internal docs
         mapped_topic = Karafka::App.config.topic_mapper.incoming(kafka_message.topic)
@@ -22,14 +22,14 @@ module Karafka
         message = Message.new(mapped_topic, kafka_message)
         controller.params = message
 
-        Karafka.monitor.notice(self.class, message)
+        Karafka.monitor.notice(self, message)
 
         controller.schedule
         # This is on purpose - see the notes for this method
         # rubocop:disable RescueException
       rescue Exception => e
         # rubocop:enable RescueException
-        Karafka.monitor.notice_error(self.class, e)
+        Karafka.monitor.notice_error(self, e)
       end
     end
   end